requirements.txt
gspread
google-auth
requests
pillow
